+++ phpunit/functional/Glpi/Dashboard/WidgetTest.php
@@ -62,7 +62,7 @@ class WidgetTest extends DbTestCase
                 'nb_series' => 4,
                 'revert'    => true,
                 'expected'  => [
-                    'names'  => [0, 1, 2, 3],
+                    'names'  => ['a', 'b', 'c', 'd'],
                     'colors' => [
                         '#a6a6a6',
                         '#808080',
@@ -75,7 +75,7 @@ class WidgetTest extends DbTestCase
                 'nb_series' => 4,
                 'revert'    => false,
                 'expected'  => [
-                    'names'  => [0, 1, 2, 3],
+                    'names'  => ['a', 'b', 'c', 'd'],
                     'colors' => [
                         '#595959',
                         '#808080',