+++ phpunit/functional/ProblemTest.php
@@ -35,6 +35,7 @@
 namespace tests\units;
 
 use DbTestCase;
+use PHPUnit\Framework\Attributes\DataProvider;
 use Problem;
 
 /* Test for inc/problem.class.php */
@@ -322,5 +323,107 @@ class ProblemTest extends DbTestCase
 
         // Assert: make sure some html was generated
         $this->assertNotEmpty($html);
+    }
+
+    public static function canAddDocumentProvider(): iterable
+    {
+        yield [
+            'profilerights' => [
+                'followup' => 0,
+                'problem'  => 0,
+                'document' => 0,
+            ],
+            'expected' => false,
+        ];
+
+        yield [
+            'profilerights' => [
+                'followup' => \ITILFollowup::ADDMYTICKET,
+                'problem'  => 0,
+                'document' => 0,
+            ],
+            'expected' => true,
+        ];
+
+        yield [
+            'profilerights' => [
+                'followup' => 0,
+                'problem'  => UPDATE,
+                'document' => 0,
+            ],
+            'expected' => false,
+        ];
+
+        yield [
+            'profilerights' => [
+                'followup' => 0,
+                'problem'  => 0,
+                'document' => CREATE,
+            ],
+            'expected' => false,
+        ];
+
+        yield [
+            'profilerights' => [
+                'followup' => \ITILFollowup::ADDMYTICKET,
+                'problem'  => UPDATE,
+                'document' => 0,
+            ],
+            'expected' => true,
+        ];
+
+        yield [
+            'profilerights' => [
+                'followup' => \ITILFollowup::ADDMYTICKET,
+                'problem'  => 0,
+                'document' => CREATE,
+            ],
+            'expected' => true,
+        ];
+
+        yield [
+            'profilerights' => [
+                'followup' => 0,
+                'problem'  => CREATE,
+                'document' => CREATE,
+            ],
+            'expected' => false,
+        ];
+    }
+
+    #[DataProvider('canAddDocumentProvider')]
+    public function testCanAddDocument(array $profilerights, bool $expected): void
+    {
+        global $DB;
+
+        foreach ($profilerights as $right => $value) {
+            $this->assertTrue($DB->update(
+                'glpi_profilerights',
+                ['rights' => $value],
+                [
+                    'profiles_id'  => 4,
+                    'name'         => $right,
+                ]
+            ));
+        }
+
+        $this->login();
+
+        $problem = $this->createItem(\Problem::class, [
+            'name' => 'Problem Test',
+            'content' => 'Problem content',
+            '_actors' => [
+                'requester' => [
+                    [
+                        'itemtype'  => 'User',
+                        'items_id'  => \Session::getLoginUserID(),
+                    ],
+                ],
+            ],
+        ]);
+
+        $input = ['itemtype' => \Problem::class, 'items_id' => $problem->getID()];
+        $doc = new \Document();
+        $this->assertEquals($expected, $doc->can(-1, CREATE, $input));
     }
 }